+++ monaco/javascript/example/index.js
@@ -30,17 +30,20 @@ const docs = buildStepDocuments(
 )
 const index = jsSearchIndex(docs)
 
-configure(monaco, index, expressions)
-
-monaco.editor.create(document.getElementById('root'), {
+const editor = monaco.editor.create(document.getElementById('root'), {
   value: `@foo
 Feature: Hello
   Scenario: Hi
     Given I have 58 cukes in my belly
     And this is an undefined step
+      | some | poorly |
+      | formatted | table |
 `,
   language: 'gherkin',
   theme: 'vs-dark',
   // semantic tokens provider is disabled by default
   'semanticHighlighting.enabled': true
-});
+})
+
+configure(monaco, editor, index, expressions)
+