+++ language-service/CHANGELOG.md
@@ -17,6 +17,25 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [0.1.1] - 2021-09-08
+
+### Fixed
+
+* Fix insertion of completion items so it always replaces the full line rather than appending to the end.
+  ([#1737](https://github.com/cucumber/common/pull/1737)
+   [aslakhellesoy])
+
+## [0.1.0] - 2021-09-07
+
+### Added
+
+* Add `CucumberInfoBuilder` and `CucumberInfo`
+  ([#1734](https://github.com/cucumber/common/pull/1734)
+   [aslakhellesoy])
+* Document Formatting
+  ([#1732](https://github.com/cucumber/common/pull/1732)
+   [aslakhellesoy])
+
 ## [0.0.1] - 2021-09-02
 
 ### Added
@@ -24,7 +43,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * First release
 
 <!-- Releases -->
-[Unreleased]: https://github.com/cucumber/common/compare/language-service/v0.0.1...main
+[Unreleased]: https://github.com/cucumber/common/compare/language-service/v0.1.1...main
+[0.1.1]:      https://github.com/cucumber/common/compare/language-service/v0.1.0...v0.1.1
+[0.1.0]:      https://github.com/cucumber/common/compare/language-service/v0.0.1...v0.1.0
 [0.0.1]:      https://github.com/cucumber/common/tree/language-service/v0.0.1
 
 <!-- Contributors in alphabetical order -->